limited seeds to 24-bit numbers
So lua uses doubles to store all numbers.
Doubles have 52 bit significands, so upto 2^52 there's integer precision.
I square rngState, meaning that what used to be up to 33 bits became up to 66.
At this point interger precision is long gone.
In rare cases when the squared value was exactly halfway between two options it had to decide which way to round it.
This meant sometimes 32 vs 64 bit (and presumably OSes) led to different sigils.
Now rngState is limited to 26 bits, which becomes 52 when squared, so this can't happen.
24 bit input is just because it's a nicer number, and because it's small enough that it's not a pain to copy all the digits.